moved func to helper package
