+++ powerscale/helper/s3_global_setting_helper.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	powerscale "dell/powerscale-go-client"
 	"terraform-provider-powerscale/client"
+	"terraform-provider-powerscale/powerscale/models"
 )
 
 // GetS3GlobalSetting gets S3 GlobalSetting.
@@ -19,4 +20,40 @@ func UpdateS3GlobalSetting(ctx context.Context, client *client.Client, GlobalSet
 	updateParam = updateParam.V10S3SettingsGlobal(GlobalSettingToUpdate)
 	_, err := updateParam.Execute()
 	return err
+}
+
+// SetGlobalSetting updates the S3 Global Setting.
+func SetGlobalSetting(ctx context.Context, client *client.Client, s3GSPlan models.S3GlobalSettingResource) (models.S3GlobalSettingResource, error) {
+	var toUpdate powerscale.V10S3SettingsGlobalSettings
+	err := ReadFromState(ctx, &s3GSPlan, &toUpdate)
+	if err != nil {
+		return models.S3GlobalSettingResource{}, err
+	}
+	err = UpdateS3GlobalSetting(ctx, client, toUpdate)
+	if err != nil {
+		return models.S3GlobalSettingResource{}, err
+	}
+	globalSettings, err := GetS3GlobalSetting(ctx, client)
+	if err != nil {
+		return models.S3GlobalSettingResource{}, err
+	}
+	var state models.S3GlobalSettingResource
+	err = CopyFieldsToNonNestedModel(ctx, globalSettings.GetSettings(), &state)
+	if err != nil {
+		return models.S3GlobalSettingResource{}, err
+	}
+	return state, nil
+}
+
+// GetGlobalSetting reads the S3 Global Setting.
+func GetGlobalSetting(ctx context.Context, client *client.Client, s3GlobalSettingState models.S3GlobalSettingResource) error {
+	globalSettings, err := GetS3GlobalSetting(ctx, client)
+	if err != nil {
+		return err
+	}
+	err = CopyFieldsToNonNestedModel(ctx, globalSettings.GetSettings(), &s3GlobalSettingState)
+	if err != nil {
+		return err
+	}
+	return nil
 }
