+++ powerscale/provider/s3_global_setting_resource.go
@@ -19,7 +19,6 @@ package provider
 
 import (
 	"context"
-	powerscale "dell/powerscale-go-client"
 	"fmt"
 	"terraform-provider-powerscale/client"
 	"terraform-provider-powerscale/powerscale/helper"
@@ -107,42 +106,6 @@ func S3GlobalSettingResourceSchema() map[string]schema.Attribute {
 	}
 }
 
-// SetGlobalSetting updates the S3 Global Setting.
-func SetGlobalSetting(ctx context.Context, client *client.Client, s3GSPlan models.S3GlobalSettingResource) (models.S3GlobalSettingResource, error) {
-	var toUpdate powerscale.V10S3SettingsGlobalSettings
-	err := helper.ReadFromState(ctx, &s3GSPlan, &toUpdate)
-	if err != nil {
-		return models.S3GlobalSettingResource{}, err
-	}
-	err = helper.UpdateS3GlobalSetting(ctx, client, toUpdate)
-	if err != nil {
-		return models.S3GlobalSettingResource{}, err
-	}
-	globalSettings, err := helper.GetS3GlobalSetting(ctx, client)
-	if err != nil {
-		return models.S3GlobalSettingResource{}, err
-	}
-	var state models.S3GlobalSettingResource
-	err = helper.CopyFieldsToNonNestedModel(ctx, globalSettings.GetSettings(), &state)
-	if err != nil {
-		return models.S3GlobalSettingResource{}, err
-	}
-	return state, nil
-}
-
-// GetGlobalSetting reads the S3 Global Setting.
-func GetGlobalSetting(ctx context.Context, client *client.Client, s3GlobalSettingState models.S3GlobalSettingResource) error {
-	globalSettings, err := helper.GetS3GlobalSetting(ctx, client)
-	if err != nil {
-		return err
-	}
-	err = helper.CopyFieldsToNonNestedModel(ctx, globalSettings.GetSettings(), &s3GlobalSettingState)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 // Create allocates the resource.
 func (r *S3GlobalSettingResource) Create(ctx context.Context, request resource.CreateRequest, response *resource.CreateResponse) {
 	tflog.Info(ctx, "Creating S3 Global Setting")
@@ -153,7 +116,7 @@ func (r *S3GlobalSettingResource) Create(ctx context.Context, request resource.C
 	if response.Diagnostics.HasError() {
 		return
 	}
-	state, err := SetGlobalSetting(ctx, r.client, s3GSPlan)
+	state, err := helper.SetGlobalSetting(ctx, r.client, s3GSPlan)
 	if err != nil {
 		response.Diagnostics.AddError("Error creating s3 global setting",
 			fmt.Sprintf("Could not create s3 global setting with error: %s", err.Error()),
@@ -178,7 +141,7 @@ func (r *S3GlobalSettingResource) Read(ctx context.Context, request resource.Rea
 		return
 	}
 
-	err := GetGlobalSetting(ctx, r.client, s3GlobalSettingState)
+	err := helper.GetGlobalSetting(ctx, r.client, s3GlobalSettingState)
 	if err != nil {
 		response.Diagnostics.AddError("Error reading s3 global setting",
 			fmt.Sprintf("Could not read s3 global setting with error: %s", err.Error()),
@@ -204,7 +167,7 @@ func (r *S3GlobalSettingResource) Update(ctx context.Context, request resource.U
 		return
 	}
 
-	state, err := SetGlobalSetting(ctx, r.client, s3GSPlan)
+	state, err := helper.SetGlobalSetting(ctx, r.client, s3GSPlan)
 	if err != nil {
 		response.Diagnostics.AddError("Error updating s3 global setting",
 			fmt.Sprintf("Could not update s3 global setting with error: %s", err.Error()),
@@ -230,7 +193,7 @@ func (r S3GlobalSettingResource) Delete(ctx context.Context, request resource.De
 // ImportState imports the resource state.
 func (r S3GlobalSettingResource) ImportState(ctx context.Context, request resource.ImportStateRequest, response *resource.ImportStateResponse) {
 	var s3GlobalSettingState models.S3GlobalSettingResource
-	err := GetGlobalSetting(ctx, r.client, s3GlobalSettingState)
+	err := helper.GetGlobalSetting(ctx, r.client, s3GlobalSettingState)
 	if err != nil {
 		response.Diagnostics.AddError("Error importing s3 global setting",
 			fmt.Sprintf("Could not import s3 global setting with error: %s", err.Error()),
